Expanded date range for imported data

+++ processed-data/import_NZ_data.m
@@ -7,13 +7,13 @@ close all
 % *count*.csv in the .gitignore 
 
 % Range of file dates (datestamp on folder/file name) needed
-fileDates = [datetime(2025, 6, 19):7:datetime(2025, 10, 9), datetime(2025, 10, 23)];    
+fileDates = [datetime(2025, 6, 5):7:datetime(2025, 10, 9), datetime(2025, 10, 23)];    
 
 % Name of local directory in which to put data files
 local_dir = "2025\";
 
 % Path to ACEFA directory in which data is stored
-ACEFA_dir = "C:\Users\mpl31\MyDocs_local\ACEFA_forecasting_repos\forecast-hub\processed-data\2025\";
+ACEFA_dir = "V:\mpl31\ACEFA_forecasting_repos\forecast-hub\processed-data\2025\";
 
 nFileDates = length(fileDates);
 
@@ -28,12 +28,23 @@ for iFileDate = 1:nFileDates
     % Specify file names for hospitalisation and case data (SARSCOV2 only)
     dates_fName = "date-information-" + dateString + ".csv";
     hosp_fName = "NZ-hospitalisations-count-" + dateString + ".csv";
+    hosp_fName_old = "SARSCOV2-NZ-hospitalisations-count-" + dateString + ".csv";
     case_fName = "SARSCOV2-all-tests-case-count-" + dateString + ".csv";
 
-    % Copy the date information and hopsitalisation data directly across (as this only contains
-    % NZ data)
+    % Copy the date information across 
     copyfile(source_dir+dates_fName, dest_dir);
-    copyfile(source_dir+hosp_fName, dest_dir);
+
+    % Copy the hopsitalisation data across (as this only contains NZ data)
+    % Try the standard file name first
+    if exist(source_dir+hosp_fName, 'file' )
+        copyfile(source_dir+hosp_fName, dest_dir);
+    % If the standard file name doesn't exist, try the old file name which
+    % was used in some of the early rounds
+    elseif exist(source_dir+hosp_fName_old, 'file' )
+        copyfile(source_dir+hosp_fName_old, dest_dir);
+    else
+        fprintf('Warning: no hopsitalisation data found for file date %s\n', dateString)
+    end
 
     % Read in the case data
     tbl = readtable(source_dir+case_fName);
